requirement.txt
mysql-connector-python
beautifulsoup4
selenium
webdriver-manager
python-dotenv
